+++ packages/components/src/components/Badge/Badge.tsx
@@ -1,13 +1,19 @@
 import type { PropsWithChildren } from "react";
 import styles from "./Badge.module.scss";
 import clsx from "clsx";
-import type { PropsWithClassName } from "@/lib/types/props";
+import {
+  alphaColors,
+  isAlphaColor,
+  type PropsWithClassName,
+} from "@/lib/types/props";
 import { type PropsContext, PropsContextProvider } from "@/lib/propsContext";
 import type { FlowComponentProps } from "@/lib/componentFactory/flowComponent";
 import { flowComponent } from "@/lib/componentFactory/flowComponent";
 import type { PressEvent } from "@react-types/shared";
 import { Button } from "@/components/Button";
 import { IconClose } from "@/components/Icon/components/icons";
+import { useLocalizedStringFormatter } from "react-aria";
+import locales from "./locales/*.locale.json";
 
 export const badgeColors = [
   "neutral",
@@ -19,8 +25,7 @@ export const badgeColors = [
   "green",
   "orange",
   "red",
-  "dark",
-  "light",
+  ...alphaColors,
 ] as const;
 export type BadgeColors = (typeof badgeColors)[number];
 
@@ -48,6 +53,8 @@ export const Badge = flowComponent("Badge", (props) => {
     isDisabled,
     ...rest
   } = props;
+
+  const stringFormatter = useLocalizedStringFormatter(locales);
 
   const rootClassName = clsx(
     styles.badge,
@@ -97,10 +104,11 @@ export const Badge = flowComponent("Badge", (props) => {
           <Button
             className={styles.close}
             size="s"
-            color={color === "light" ? "light" : "dark"}
+            color={isAlphaColor(color) ? color : "dark"}
             variant="plain"
             onPress={onClose}
             isDisabled={isDisabled}
+            aria-label={stringFormatter.format("badge.remove")}
           >
             <IconClose />
           </Button>