+++ packages/components/src/components/Calendar/components/RangeCalendar/view.ts
@@ -1,0 +1,10 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+import type { RangeCalendar } from "./RangeCalendar";
+import type { ViewComponent } from "@/lib/viewComponentContext";
+
+declare global {
+  interface FlowViewComponents {
+    RangeCalendar: ViewComponent<typeof RangeCalendar>;
+  }
+}