+++ packages/components/src/components/List/components/Header/components/ActiveFilters/ActiveFilters.tsx
@@ -1,5 +1,4 @@
 import type { FC } from "react";
-import React from "react";
 import { useList } from "@/components/List/hooks/useList";
 import styles from "./ActiveFilters.module.scss";
 import locales from "../../../../locales/*.locale.json";
@@ -18,43 +17,60 @@ import DivView from "@/views/DivView";
 import BadgeView from "@/views/BadgeView";
 import TooltipTriggerView from "@/views/TooltipTriggerView";
 import TextView from "@/views/TextView";
+import { Filter } from "@/components/List/model/filter/Filter";
+import { transformDateValueToFormattedDate } from "@/lib/date/transformDateValueToFormattedDate";
+import { DateRangeFilter } from "@/components/List/model/filter/DateRangeFilter";
 
 export const ActiveFilters: FC = observer(() => {
   const list = useList();
   const formatter = useLocalizedStringFormatter(locales);
 
-  const activeFilterValues = list.filters
-    .flatMap((f) => f.values)
-    .filter((v) => v.isActive);
+  const activeFilters = list.filters.flatMap((f) => {
+    if (f instanceof DateRangeFilter) {
+      const value = f.getValue();
+      if (value) {
+        return [
+          <BadgeView key={f.name} onClose={() => f.clear()}>
+            <TextView>
+              {`${transformDateValueToFormattedDate(value.start)} - ${transformDateValueToFormattedDate(value.end)}`}
+            </TextView>
+          </BadgeView>,
+        ];
+      }
+    }
 
-  const activeFilters = activeFilterValues.map((v) => (
-    <BadgeView key={v.id} onClose={() => v.deactivate()}>
-      <TextView>{v.render()}</TextView>
-    </BadgeView>
-  ));
+    return f.values
+      .filter((v) => v.isActive)
+      .map((v) => (
+        <BadgeView key={v.id} onClose={() => v.deactivate()}>
+          <TextView>{v.render()}</TextView>
+        </BadgeView>
+      ));
+  });
 
-  const someFiltersChanged =
-    list.filters.filter((f) => f.hasChanged()).length > 0;
+  const storingAvailable = list.filters.some((f) => f.isStoringAvailable());
+  const hasChanges = list.filters.some((f) => f.hasChanges());
 
-  const storeFiltersButton = list.supportsSettingsStorage &&
-    someFiltersChanged && (
-      <TooltipTriggerView>
-        <TooltipView>
-          <Translate locales={locales}>list.filters.store</Translate>
-        </TooltipView>
-        <ButtonView
-          size="s"
-          variant="plain"
-          color="secondary"
-          onPress={() => list.storeFilterDefaultSettings()}
-          aria-label={formatter.format("list.filters.store")}
-        >
-          <IconSave />
-        </ButtonView>
-      </TooltipTriggerView>
-    );
+  const storeFiltersButton = storingAvailable && hasChanges && (
+    <TooltipTriggerView>
+      <TooltipView>
+        <Translate locales={locales}>list.filters.store</Translate>
+      </TooltipView>
+      <ButtonView
+        size="s"
+        variant="plain"
+        color="secondary"
+        onPress={() =>
+          Filter.storeFilters(list, { autosave: false, manualSave: true })
+        }
+        aria-label={formatter.format("list.filters.store")}
+      >
+        <IconSave />
+      </ButtonView>
+    </TooltipTriggerView>
+  );
 
-  const resetFiltersButton = someFiltersChanged ? (
+  const resetFiltersButton = hasChanges ? (
     <TooltipTrigger>
       <TooltipView>
         <Translate locales={locales}>list.filters.reset</Translate>