+++ packages/components/src/components/List/model/ReactTable.ts
@@ -23,7 +23,8 @@ import type {
 } from "@/components/List/model/types";
 import type { SearchValue } from "@/components/List/model/search/types";
 import type { Dispatch, SetStateAction } from "react";
-import { useEffect, useState } from "react";
+import { useEffect, useEffectEvent, useState } from "react";
+import { Filter } from "./filter/Filter";
 
 export class ReactTable<T, TMeta = unknown> {
   public readonly list: List<T, TMeta>;
@@ -39,7 +40,7 @@ export class ReactTable<T, TMeta = unknown> {
     this.list = list;
 
     const defaultSorting = this.list.sorting.filter(
-      (s) => s.defaultEnabled !== false,
+      (s) => s.initialEnabled !== false,
     );
 
     const [sortingState, updateSortingState] = useState<ColumnSort[]>(
@@ -107,12 +108,25 @@ export class ReactTable<T, TMeta = unknown> {
       ...tableOptions,
     });
 
-    useEffect(() => {
+    const reactTableState = table.getState();
+
+    const onFiltersChanged = useEffectEvent(() => {
+      Filter.storeFilters(this.list, { autosave: true });
+    });
+
+    const onStateChanged = useEffectEvent(() => {
       if (onChange) {
         onChange(this.list);
       }
-      this.list.autosaveFilterSettings();
-    }, [this.list, onChange, table]);
+    });
+
+    useEffect(() => {
+      onStateChanged();
+    }, [reactTableState]);
+
+    useEffect(() => {
+      onFiltersChanged();
+    }, [reactTableState.columnFilters]);
 
     return table;
   }
@@ -124,7 +138,7 @@ export class ReactTable<T, TMeta = unknown> {
     const additionalHiddenSorting = this.list.sorting
       .filter(
         (s) =>
-          s.defaultEnabled === "hidden" &&
+          s.initialEnabled === "hidden" &&
           !newSortingState.some((existing) => existing.id === s.property),
       )
       .map((s) => s.getReactTableColumnSort());