+++ packages/components/src/components/List/stories/Default.stories.tsx
@@ -31,6 +31,8 @@ import { TableRow } from "@/components/List/setupComponents/TableRow";
 import { TableCell } from "@/components/List/setupComponents/TableCell";
 import SkeletonText from "@/components/SkeletonText";
 import Skeleton from "@/components/Skeleton";
+import { DateTime } from "luxon";
+import { getLocalTimeZone, today } from "@internationalized/date";
 
 const loadDomains: AsyncDataLoader<Domain> = async (opts) => {
   const response = await getDomains({
@@ -85,7 +87,9 @@ const meta: Meta<typeof List> = {
               values={availableTypes}
               property="type"
               name="Type"
-              initialSelected={["Subdomain"]}
+              autosave
+              manualSave
+              defaultSelected={["Domain"]}
             />
 
             <DomainList.Search autoFocus />
@@ -413,6 +417,43 @@ export const WithSecondaryFilters: Story = {
           </DomainList.List>
         </Section>
       </SettingsProvider>
+    );
+  },
+};
+
+export const WithDateRangeFilter: Story = {
+  render: () => {
+    const List = typedList<{
+      id: string;
+      date: string;
+    }>();
+
+    return (
+      <List.List batchSize={5} aria-label="Invoices">
+        <List.StaticData
+          data={[
+            { id: "RG100000", date: DateTime.now().toISO() },
+            { id: "RG100001", date: DateTime.now().minus({ day: 7 }).toISO() },
+            { id: "RG100002", date: DateTime.now().minus({ day: 14 }).toISO() },
+          ]}
+        />
+        <List.Filter
+          property="date"
+          name="Date"
+          mode="dateRange"
+          dateRangeOptions={{ maxValue: today(getLocalTimeZone()) }}
+        />
+        <List.Item textValue={(invoice) => invoice.id}>
+          {(invoice) => (
+            <ListItemView>
+              <Heading>{invoice.id}</Heading>
+              <Text>
+                {DateTime.fromISO(invoice.date).toFormat("dd.MM.yyyy")}
+              </Text>
+            </ListItemView>
+          )}
+        </List.Item>
+      </List.List>
     );
   },
 };