+++ packages/components/src/integrations/react-hook-form/components/Field/Field.tsx
@@ -1,6 +1,9 @@
 import { useFormContext } from "@/integrations/react-hook-form/components/FormContextProvider/FormContextProvider";
-import { dynamic, type PropsContext } from "@/lib/propsContext";
-import { PropsContextProvider } from "@/lib/propsContext";
+import {
+  dynamic,
+  type PropsContext,
+  PropsContextProvider,
+} from "@/lib/propsContext";
 import { type PropsWithChildren } from "react";
 import {
   type ControllerProps,
@@ -26,6 +29,37 @@ export function Field<T extends FieldValues>(props: FieldProps<T>) {
     ...props,
     rules: {
       ...props.rules,
+      min:
+        typeof rest.rules?.min === "number"
+          ? {
+              value: rest.rules.min,
+              message: stringFormatter.format("min", {
+                number: rest.rules.min,
+              }),
+            }
+          : rest.rules?.min,
+      max:
+        typeof rest.rules?.max === "number"
+          ? {
+              value: rest.rules.max,
+              message: stringFormatter.format("max", {
+                number: rest.rules.max,
+              }),
+            }
+          : rest.rules?.max,
+      required:
+        typeof rest.rules?.required === "boolean"
+          ? stringFormatter.format("required")
+          : rest.rules?.required,
+      pattern:
+        rest.rules?.pattern instanceof RegExp
+          ? {
+              value: rest.rules.pattern,
+              message: stringFormatter.format("pattern", {
+                pattern: rest.rules.pattern.source,
+              }),
+            }
+          : rest.rules?.pattern,
       minLength:
         typeof rest.rules?.minLength === "number"
           ? {
@@ -100,6 +134,7 @@ export function Field<T extends FieldValues>(props: FieldProps<T>) {
     TextField: fieldProps,
     TextArea: fieldProps,
     MarkdownEditor: fieldProps,
+    CodeEditor: fieldProps,
     Checkbox: {
       ...fieldProps,
       isSelected: value,