+++ packages/components/src/lib/remote/types.ts
@@ -1,5 +1,5 @@
 export type FieldProps<T, P> = {
   value?: T;
   defaultValue?: T;
-  onChange?: (value: T) => void;
+  onChange?: (value: T, ...args: never[]) => void;
 } & P;